LinkMap S7 container class

LinkMap is an S7 class to organize and manage multiple sets of factors, for instance when tracing or converting feature IDs across databases. Methods for LinkMap aim to follow factor behaviour.

Usage

LinkMap(x)

Arguments

x

a data.frame with two named columns that can be coerced to factors and an optional third column of values associated with those two columns.

Value

a LinkMap object.

Slots

levels

Named list of character vectors depicting levels.

value

Numeric vector or boolean scalar.

is_bool

Boolean, returns TRUE if this object contains no value.

See also

Examples

# Generate random linkage input
x <- data.frame(
    a = sample(letters[seq(3)], 10, replace = TRUE),
    A = sample(LETTERS[seq(3)], 10, replace = TRUE)
)

# Create LinkMap
LinkMap(x)
#>    a A
#> 1  c C
#> 2  a C
#> 3  b A
#> 4  c B
#> 5  c B
#> 6  c C
#> 7  b C
#> 8  a B
#> 9  a C
#> 10 a A