Generate a random MultiFactor or LinkMap

Randomly generate a valid MultiFactor or LinkMap object.

called by randomMultiFactor, shouldn't be called by user.

Usage

randomMultiFactor(n_types = 6, n_features = 100, sparseness = 0.5)

randomLinkMap(x = list(lower = letters, UPPER = LETTERS), sparseness = 0.5)

Arguments

n_types

Numeric scalar, number of types of features to generate

n_features

Numeric scalar, number of features per type

sparseness

Numeric scalar, proportion: How rare are connections. Default is 0.5.

x

optional list of two named vectors of features to use. Default is list(lower = letters, UPPER = LETTERS).

Value

a randomly generated object of the specified class.

See also

Examples

# Make a random MultiFactor object
randomMultiFactor()
#> A MultiFactor::MultiFactor list S7_object,
#>     6 feature types across 5 LinkMaps.
#> 
#>       a   b   c   d   e   f
#> a2b 100 100   .   .   .   .
#> b2c   . 100 100   .   .   .
#> c2d   .   . 100 100   .   .
#> d2e   .   .   . 100 100   .
#> e2f   .   .   .   . 100 100
#> 
#> Values represent unique feature names in that LinkMap.
#> 
#> Levels:
#> a : 100 Levels: a_001 a_002 ... a_100 
#> b : 100 Levels: b_001 b_002 ... b_100 
#> c : 100 Levels: c_001 c_002 ... c_100 
#> d : 100 Levels: d_001 d_002 ... d_100 
#> e : 100 Levels: e_001 e_002 ... e_100 
#> f : 100 Levels: f_001 f_002 ... f_100 

# Make a random LinkMap object
randomLinkMap()
#>     lower UPPER
#> 2       b     A
#> 4       d     A
#> 7       g     A
#> 11      k     A
#> 13      m     A
#> 21      u     A
#> 22      v     A
#> 23      w     A
#> 24      x     A
#> 25      y     A
#> 26      z     A
#> 27      a     B
#> 29      c     B
#> 31      e     B
#> 32      f     B
#> 34      h     B
#> 37      k     B
#> 39      m     B
#> 42      p     B
#> 43      q     B
#> 44      r     B
#> 47      u     B
#> 49      w     B
#> 53      a     C
#> 57      e     C
#> 59      g     C
#> 61      i     C
#> 65      m     C
#> 68      p     C
#> 70      r     C
#> 75      w     C
#> 79      a     D
#> 80      b     D
#> 82      d     D
#> 85      g     D
#> 86      h     D
#> 90      l     D
#> 96      r     D
#> 97      s     D
#> 98      t     D
#> 99      u     D
#> 101     w     D
#> 102     x     D
#> 103     y     D
#> 105     a     E
#> 106     b     E
#> 108     d     E
#> 109     e     E
#> 110     f     E
#> 111     g     E
#> 113     i     E
#> 114     j     E
#> 116     l     E
#> 117     m     E
#> 121     q     E
#> 122     r     E
#> 123     s     E
#> 124     t     E
#> 125     u     E
#> 128     x     E
#> 129     y     E
#> 138     h     F
#> 139     i     F
#> 142     l     F
#> 144     n     F
#> 145     o     F
#> 149     s     F
#> 151     u     F
#> 152     v     F
#> 153     w     F
#> 154     x     F
#> 155     y     F
#> 156     z     F
#> 158     b     G
#> 160     d     G
#> 161     e     G
#> 162     f     G
#> 165     i     G
#> 166     j     G
#> 170     n     G
#> 171     o     G
#> 172     p     G
#> 173     q     G
#> 176     t     G
#> 179     w     G
#> 180     x     G
#> 184     b     H
#> 186     d     H
#> 189     g     H
#> 191     i     H
#> 192     j     H
#> 194     l     H
#> 197     o     H
#> 198     p     H
#> 199     q     H
#> 200     r     H
#> 201     s     H
#> 202     t     H
#> 203     u     H
#> 204     v     H
#> 205     w     H
#> 207     y     H
#> 208     z     H
#> 209     a     I
#> 210     b     I
#> 211     c     I
#> 213     e     I
#> 214     f     I
#> 217     i     I
#> 218     j     I
#> 219     k     I
#> 223     o     I
#> 224     p     I
#> 225     q     I
#> 227     s     I
#> 231     w     I
#> 232     x     I
#> 234     z     I
#> 235     a     J
#> 236     b     J
#> 237     c     J
#> 243     i     J
#> 248     n     J
#> 251     q     J
#> 255     u     J
#> 260     z     J
#> 261     a     K
#> 263     c     K
#> 264     d     K
#> 265     e     K
#> 266     f     K
#> 268     h     K
#> 273     m     K
#> 274     n     K
#> 275     o     K
#> 276     p     K
#> 278     r     K
#> 280     t     K
#> 281     u     K
#> 282     v     K
#> 283     w     K
#> 285     y     K
#> 286     z     K
#> 287     a     L
#> 290     d     L
#> 291     e     L
#> 293     g     L
#> 294     h     L
#> 296     j     L
#> 297     k     L
#> 302     p     L
#> 304     r     L
#> 307     u     L
#> 312     z     L
#> 313     a     M
#> 314     b     M
#> 318     f     M
#> 320     h     M
#> 321     i     M
#> 323     k     M
#> 325     m     M
#> 327     o     M
#> 328     p     M
#> 331     s     M
#> 333     u     M
#> 334     v     M
#> 336     x     M
#> 337     y     M
#> 339     a     N
#> 340     b     N
#> 343     e     N
#> 346     h     N
#> 347     i     N
#> 349     k     N
#> 353     o     N
#> 357     s     N
#> 360     v     N
#> 361     w     N
#> 363     y     N
#> 365     a     O
#> 367     c     O
#> 370     f     O
#> 371     g     O
#> 373     i     O
#> 374     j     O
#> 377     m     O
#> 378     n     O
#> 380     p     O
#> 381     q     O
#> 382     r     O
#> 383     s     O
#> 384     t     O
#> 388     x     O
#> 389     y     O
#> 390     z     O
#> 391     a     P
#> 392     b     P
#> 393     c     P
#> 394     d     P
#> 395     e     P
#> 396     f     P
#> 399     i     P
#> 402     l     P
#> 403     m     P
#> 404     n     P
#> 405     o     P
#> 407     q     P
#> 408     r     P
#> 409     s     P
#> 410     t     P
#> 411     u     P
#> 415     y     P
#> 416     z     P
#> 417     a     Q
#> 418     b     Q
#> 419     c     Q
#> 420     d     Q
#> 421     e     Q
#> 422     f     Q
#> 423     g     Q
#> 424     h     Q
#> 425     i     Q
#> 427     k     Q
#> 431     o     Q
#> 435     s     Q
#> 437     u     Q
#> 438     v     Q
#> 441     y     Q
#> 442     z     Q
#> 444     b     R
#> 445     c     R
#> 447     e     R
#> 450     h     R
#> 454     l     R
#> 458     p     R
#> 459     q     R
#> 460     r     R
#> 461     s     R
#> 465     w     R
#> 466     x     R
#> 467     y     R
#> 469     a     S
#> 473     e     S
#> 474     f     S
#> 475     g     S
#> 476     h     S
#> 477     i     S
#> 478     j     S
#> 479     k     S
#> 483     o     S
#> 484     p     S
#> 485     q     S
#> 486     r     S
#> 487     s     S
#> 489     u     S
#> 490     v     S
#> 494     z     S
#> 501     g     T
#> 503     i     T
#> 504     j     T
#> 505     k     T
#> 506     l     T
#> 509     o     T
#> 512     r     T
#> 514     t     T
#> 515     u     T
#> 518     x     T
#> 521     a     U
#> 525     e     U
#> 526     f     U
#> 527     g     U
#> 531     k     U
#> 532     l     U
#> 537     q     U
#> 538     r     U
#> 539     s     U
#> 540     t     U
#> 543     w     U
#> 545     y     U
#> 546     z     U
#> 547     a     V
#> 548     b     V
#> 550     d     V
#> 552     f     V
#> 553     g     V
#> 554     h     V
#> 556     j     V
#> 557     k     V
#> 558     l     V
#> 559     m     V
#> 562     p     V
#> 564     r     V
#> 566     t     V
#> 567     u     V
#> 568     v     V
#> 574     b     W
#> 575     c     W
#> 576     d     W
#> 577     e     W
#> 579     g     W
#> 582     j     W
#> 583     k     W
#> 591     s     W
#> 592     t     W
#> 599     a     X
#> 602     d     X
#> 603     e     X
#> 604     f     X
#> 605     g     X
#> 607     i     X
#> 610     l     X
#> 616     r     X
#> 617     s     X
#> 619     u     X
#> 620     v     X
#> 621     w     X
#> 624     z     X
#> 625     a     Y
#> 626     b     Y
#> 628     d     Y
#> 630     f     Y
#> 635     k     Y
#> 636     l     Y
#> 643     s     Y
#> 644     t     Y
#> 645     u     Y
#> 646     v     Y
#> 649     y     Y
#> 651     a     Z
#> 653     c     Z
#> 654     d     Z
#> 655     e     Z
#> 656     f     Z
#> 660     j     Z
#> 663     m     Z
#> 664     n     Z
#> 673     w     Z
#> 675     y     Z